\name{PTMsToPathwaysData-package}
\alias{PTMsToPathwaysData-package}
\docType{package}
\title{ExperimentHub Data for PTMsToPathways}
\description{
PTMsToPathwaysData is a companion ExperimentHub data package for PTMsToPathways.
It provides access to large post-translational modification data resources.
}
\details{
Package resources are distributed through ExperimentHub.
Use \code{PTMsToPathwaysData()} to retrieve records tagged for this package.
}
\author{
Lucia Williams
}
\seealso{
\code{\link{PTMsToPathwaysData}},
\code{\link[ExperimentHub]{ExperimentHub}}
}
\keyword{package}
