% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{PTMsToPathwaysData-package}
\alias{PTMsToPathwaysData-package}
\title{PTMsToPathwaysData: ExperimentHub Data for PTMsToPathways}
\description{
Companion ExperimentHub data package for PTMsToPathways. It provides raw,
intermediate, and derived data resources used in the PTMsToPathways
vignettes and examples.
}
\details{
Available data resources include:

Getting started vignette:
\itemize{
\item \code{AlldataPTMs.txt} and \code{ex_full_ptm_table.rda}
}

Raw data processing vignette:
\itemize{
\item \code{phospho_cleaned_mapped.txt} and \code{newphos.rda}
}

Cytoscape graphing vignette:
\itemize{
\item \code{FunctionKey.txt} and \code{function_key.rda}
}

BRCA Networks vignette:
\itemize{
\item \code{bioplanet_pathway_June2025.csv} and \code{bioplanet_pathways.rda}
\item \code{PhosphoDiscoData_mmc9.txt}
\item \code{brca_clusterlist_data.rda}
\item \code{brca_CCCN_data.rda}
\item \code{BRCA_stringdb.edges.rda}
\item \code{BRCA_genemania.edges.rda}
\item \code{BRCA_PCN.data.rda}
}

Use \code{\link[=PTMsToPathwaysData]{PTMsToPathwaysData()}} to query the ExperimentHub records registered by
this package.
}
\examples{
\donttest{
eh <- ExperimentHub::ExperimentHub()
query(eh, "PTMsToPathwaysData")
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UM-Applied-Algorithms-Lab/PTMsToPathwaysData}
  \item Report bugs at \url{https://github.com/UM-Applied-Algorithms-Lab/PTMsToPathwaysData/issues}
}

}
\author{
\strong{Maintainer}: Lucia Williams \email{lucia.williams@umt.edu} (\href{https://orcid.org/0000-0003-3785-0247}{ORCID})

Authors:
\itemize{
  \item Lucia Williams \email{lucia.williams@umt.edu} (\href{https://orcid.org/0000-0003-3785-0247}{ORCID})
  \item Mark Grimes (\href{https://orcid.org/0000-0003-2673-5892}{ORCID})
}

}
\keyword{package}
