% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTMsToPathwaysData.R
\name{PTMsToPathwaysData}
\alias{PTMsToPathwaysData}
\title{Query PTMsToPathways Data Resources in ExperimentHub}
\usage{
PTMsToPathwaysData()
}
\value{
A subsetted \code{ExperimentHub} object.
}
\description{
Returns ExperimentHub records tagged for PTMsToPathways BRCA data.
}
\examples{
 ptm_data <- PTMsToPathwaysData()
 ptm_data
}
